some changes

+++ services/tasks.py
@@ -1,20 +1,21 @@
-from fastapi import Depends
-from dependensy import get_db
+from fastapi import HTTPException,status
 from shemas.dbmodels import Task
 from shemas.models import TaskModels
 import shemas.dbmodels as dbmodels
 from sqlalchemy.orm import Session
 
-def get_tasks_server(user_id:int,db: Session = Depends(get_db)): 
+def get_tasks_server(user_id:int,db: Session): 
     return db.query(dbmodels.Task).filter(Task.user_id==user_id).all()
 
-def add_to_db_server(task:TaskModels,user_id:int,db: Session=Depends(get_db)):
+def add_to_db_server(task:TaskModels,user_id:int,db: Session):
     db.add(Task(**task.model_dump(),user_id=user_id))
     db.commit()
     return "Sucsses"
 
-def update_server(id:int,task:TaskModels,user_id:int,db: Session=Depends(get_db)):
+def update_server(id:int,task:TaskModels,user_id:int,db: Session):
     db_task=db.query(dbmodels.Task).filter(dbmodels.Task.id==id).first()
+    if db_task ==None:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,detail="Task not found")
     if db_task.user_id==user_id:
         db_task.name=task.name
         db_task.description=task.description
@@ -22,18 +23,15 @@ def update_server(id:int,task:TaskModels,user_id:int,db: Session=Depends(get_db)
         db.commit()
         return db.query(dbmodels.Task).filter(dbmodels.Task.id==id).first()
     else:
-        return "Wrong id"
+        raise HTTPException(status_code=403, detail="Forbidden")
     
-def delete_from_db_server(id: int,user_id:int,db: Session=Depends(get_db)):
+def delete_from_db_server(id: int,user_id:int,db: Session):
     db_deleted=db.query(dbmodels.Task).filter(dbmodels.Task.id==id).first()
+    if db_deleted ==None:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,detail="Task not found")
     if db_deleted.user_id==user_id:
         db.delete(db_deleted)
         db.commit()
+        return"Sucsses"
     else:
-        return"Wrong id"
-    return"Sucsses"
-
-def delete_from_db_all_server(user_id:int,db: Session=Depends(get_db)):
-    db.query(dbmodels.Task).delete()
-    db.commit()
-    return "Sucsses"
+        raise HTTPException(status_code=403, detail="Forbidden")
